+++ src/_locales/zh-TW/menus.ts
@@ -5,6 +5,7 @@ export const locale: typeof _locale = {
   baidu_search: '百度搜尋',
   bing_dict: 'Bing 字典',
   bing_search: 'Bing 搜尋',
+  caiyuntrs: '彩雲小譯網頁翻譯',
   cambridge: '劍橋字典',
   copy_pdf_url: '複製PDF連結到剪貼簿',
   dictcn: '海詞字典',
@@ -26,6 +27,7 @@ export const locale: typeof _locale = {
   notification_youdao_err:
     '有道網頁翻譯2.0 下載後無回應，\n可能是套件無權造訪該網站，\n如果下載成功後，請忽略本訊息。',
   oxford: '牛津字典',
+  page_permission_err: '沙拉查詞「{{name}}」無權訪問此頁面。',
   page_translations: '網頁翻譯',
   saladict: '沙拉查詞',
   saladict_standalone: '沙拉查詞獨立視窗',