+++ src/background/index.ts
@@ -4,13 +4,14 @@ import {
   createActiveProfileStream,
   createProfileIDListStream
 } from '@/_helpers/profile-manager'
-import { reportGA } from '@/_helpers/analytics'
+import { reportPaveview } from '@/_helpers/analytics'
 import { message } from '@/_helpers/browser-api'
 import { startSyncServiceInterval } from './sync-manager'
 import { init as initPdf } from './pdf-sniffer'
 import { ContextMenus } from './context-menus'
 import { BackgroundServer } from './server'
 import { initBadge } from './badge'
+import { setupCaiyunTrsBackend } from './page-translate/caiyun'
 import './types'
 
 // init first to recevice self messaging
@@ -21,10 +22,12 @@ startSyncServiceInterval()
 ContextMenus.init()
 BackgroundServer.init()
 
+setupCaiyunTrsBackend()
+
 getConfig().then(async config => {
   window.appConfig = config
   initPdf(config)
-  reportGA('/background')
+  reportPaveview('/background')
   initBadge()
 
   addConfigListener(({ newConfig }) => {