+++ src/options/components/Entries/ContextMenus/index.tsx
@@ -1,9 +1,8 @@
 import React, { FC, useState, useLayoutEffect } from 'react'
 import { Row, Col } from 'antd'
-import { isFirefox } from '@/_helpers/saladict'
 import { useTranslate } from '@/_helpers/i18n'
 import { useSelector } from '@/content/redux'
-import { SortableList, arrayMove } from '@/options/components/SortableList'
+import { SortableList, reorder } from '@/options/components/SortableList'
 import { getConfigPath } from '@/options/helpers/path-joiner'
 import { useListLayout } from '@/options/helpers/layout'
 import { useUpload } from '@/options/helpers/upload'
@@ -31,21 +30,13 @@ export const ContextMenus: FC = () => {
         <SortableList
           title={t('nav.ContextMenus')}
           description={<p>{t('config.opt.contextMenus_description')}</p>}
-          list={selectedMenus
-            .filter(id => {
-              // FF policy
-              if (isFirefox && id === 'youdao_page_translate') {
-                return false
-              }
-              return true
-            })
-            .map(id => {
-              const item = contextMenus.all[id]
-              return {
-                value: id,
-                title: typeof item === 'string' ? t(`menus:${id}`) : item.name
-              }
-            })}
+          list={selectedMenus.map(id => {
+            const item = contextMenus.all[id]
+            return {
+              value: id,
+              title: typeof item === 'string' ? t(`menus:${id}`) : item.name
+            }
+          })}
           disableEdit={(index, item) => contextMenus.all[item.value] === 'x'}
           onAdd={() => setShowAddModal(true)}
           onEdit={index => {
@@ -59,11 +50,8 @@ export const ContextMenus: FC = () => {
             })
             setSelectedMenus(newList)
           }}
-          onSortEnd={({ oldIndex, newIndex }) => {
-            if (oldIndex === newIndex) {
-              return
-            }
-            const newList = arrayMove(selectedMenus.slice(), oldIndex, newIndex)
+          onOrderChanged={(oldIndex, newIndex) => {
+            const newList = reorder(selectedMenus, oldIndex, newIndex)
             upload({
               [getConfigPath('contextMenus', 'selected')]: newList
             })