+++ src/options/components/SortableList/index.tsx
@@ -1,10 +1,5 @@
 import React from 'react'
-import {
-  SortableContainer,
-  SortableHandle,
-  SortableElement,
-  SortEnd as _SortEnd
-} from 'react-sortable-hoc'
+import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd'
 import { List, Radio, Button, Card } from 'antd'
 import {
   PlusOutlined,
@@ -18,7 +13,7 @@ import { useTranslate } from '@/_helpers/i18n'
 
 import './_style.scss'
 
-export type SortEnd = _SortEnd
+export { reorder } from './reorder'
 
 export type SortableListItem = { value: string; title: React.ReactNode }
 
@@ -44,73 +39,8 @@ export interface SortableListProps
   onAdd?: () => void
   /** Title being selected */
   onSelect?: (e: RadioChangeEvent) => void
-  onSortEnd?: (end: SortEnd) => void
+  onOrderChanged?: (oldIndex: number, newIndex: number) => void
 }
-
-const DragHandle = React.memo(
-  SortableHandle(() => {
-    const { t } = useTranslate('common')
-    return (
-      <SwapOutlined rotate={90} title={t('sort')} style={{ cursor: 'move' }} />
-    )
-  })
-)
-
-const ProfileListItem = SortableElement(
-  ({
-    selected,
-    item,
-    disableEdit,
-    onEdit,
-    onDelete,
-    indexCopy
-  }: SortableListItemProps) => {
-    const { t } = useTranslate('options')
-
-    return (
-      <List.Item>
-        <div className="sortable-list-item">
-          {selected == null ? (
-            item.title
-          ) : (
-            <Radio value={item.value}>{item.title}</Radio>
-          )}
-          <div>
-            <DragHandle />
-            <Button
-              className="sortable-list-item-btn"
-              title={t('common:edit')}
-              shape="circle"
-              size="small"
-              icon={<EditOutlined />}
-              disabled={disableEdit != null && disableEdit(indexCopy, item)}
-              onClick={onEdit && (() => onEdit(indexCopy, item))}
-            />
-            <Button
-              title={t('common:delete')}
-              className="sortable-list-item-btn"
-              shape="circle"
-              size="small"
-              icon={<CloseOutlined />}
-              disabled={selected != null && item.value === selected}
-              onClick={onDelete && (() => onDelete(indexCopy, item))}
-            />
-          </div>
-        </div>
-      </List.Item>
-    )
-  }
-)
-
-const SortableListContainer = SortableContainer((props: SortableListProps) => (
-  <List
-    size="large"
-    dataSource={props.list}
-    renderItem={(item: any, index: number) => (
-      <ProfileListItem {...props} item={item} index={index} indexCopy={index} />
-    )}
-  />
-))
 
 export function SortableList(props: SortableListProps) {
   const { t } = useTranslate('common')
@@ -131,7 +61,93 @@ export function SortableList(props: SortableListProps) {
         value={props.selected}
         onChange={props.onSelect}
       >
-        <SortableListContainer useDragHandle {...props} />
+        <DragDropContext
+          onDragEnd={result => {
+            if (
+              props.onOrderChanged &&
+              result.destination &&
+              result.source.index !== result.destination.index
+            ) {
+              props.onOrderChanged(
+                result.source.index,
+                result.destination.index
+              )
+            }
+          }}
+        >
+          <Droppable droppableId="droppable">
+            {({ innerRef: droppableRef, droppableProps, placeholder }) => (
+              <div ref={droppableRef} {...droppableProps}>
+                <List size="large">
+                  {props.list.map((item, index) => (
+                    <Draggable
+                      key={item.value}
+                      draggableId={item.value}
+                      index={index}
+                      disableInteractiveElementBlocking
+                    >
+                      {({
+                        innerRef: draggableRef,
+                        draggableProps,
+                        dragHandleProps
+                      }) => (
+                        <div ref={draggableRef} {...draggableProps}>
+                          <List.Item key={item.value}>
+                            <div className="sortable-list-item">
+                              {props.selected == null ? (
+                                item.title
+                              ) : (
+                                <Radio value={item.value}>{item.title}</Radio>
+                              )}
+                              <div className="sortable-list-item-btns">
+                                <SwapOutlined
+                                  rotate={90}
+                                  title={t('sort')}
+                                  style={{ cursor: 'move' }}
+                                  {...dragHandleProps}
+                                />
+                                <Button
+                                  className="sortable-list-item-btn"
+                                  title={t('edit')}
+                                  shape="circle"
+                                  size="small"
+                                  icon={<EditOutlined />}
+                                  disabled={
+                                    props.disableEdit != null &&
+                                    props.disableEdit(index, item)
+                                  }
+                                  onClick={() =>
+                                    props.onEdit && props.onEdit(index, item)
+                                  }
+                                />
+                                <Button
+                                  title={t('delete')}
+                                  className="sortable-list-item-btn"
+                                  shape="circle"
+                                  size="small"
+                                  icon={<CloseOutlined />}
+                                  disabled={
+                                    props.selected != null &&
+                                    item.value === props.selected
+                                  }
+                                  onClick={() =>
+                                    props.onDelete &&
+                                    props.onDelete(index, item)
+                                  }
+                                />
+                              </div>
+                            </div>
+                          </List.Item>
+                        </div>
+                      )}
+                    </Draggable>
+                  ))}
+                  {placeholder}
+                </List>
+              </div>
+            )}
+          </Droppable>
+        </DragDropContext>
       </Radio.Group>
       {(props.isShowAdd == null || props.isShowAdd) && (
         <Button type="dashed" style={{ width: '100%' }} onClick={props.onAdd}>
@@ -140,16 +156,4 @@ export function SortableList(props: SortableListProps) {
       )}
     </Card>
   )
-}
-
-/**
- * Changes the contents of an array by moving an element to a different position
- */
-export function arrayMove<T extends any[]>(
-  arr: T,
-  from: number,
-  to: number
-): T {
-  arr.splice(to < 0 ? arr.length + to : to, 0, arr.splice(from, 1)[0])
-  return arr
 }