+++ src/options/components/SortableList/reorder.ts
@@ -1,0 +1,10 @@
+export function reorder<T extends readonly any[]>(
+  list: T,
+  startIndex: number,
+  endIndex: number
+) {
+  const result = Array.from(list)
+  const [removed] = result.splice(startIndex, 1)
+  result.splice(endIndex, 0, removed)
+  return result
+}