+++ src/popup/index.tsx
@@ -5,7 +5,7 @@ import React, { FC } from 'react'
 import ReactDOM from 'react-dom'
 import { Helmet } from 'react-helmet'
 import { AppConfig } from '@/app-config'
-import { reportGA } from '@/_helpers/analytics'
+import { reportPaveview } from '@/_helpers/analytics'
 import { getConfig } from '@/_helpers/config-manager'
 import { message, openURL } from '@/_helpers/browser-api'
 import { saveWord, Word } from '@/_helpers/record-manager'
@@ -51,14 +51,16 @@ getConfig().then(config => {
       message.send({ type: 'OPEN_QS_PANEL' })
       break
     default:
-      sendContextMenusClick(config.baOpen)
+      sendContextMenusClick(config.baOpen).then(() => {
+        window.close()
+      })
       break
   }
 })
 
 async function showPanel(config: AppConfig) {
   if (config.analytics) {
-    reportGA('/popup')
+    reportPaveview('/popup')
   }
 
   const store = await createStore()
@@ -130,28 +132,30 @@ async function sendContextMenusClick(menuItemId: string) {
   const payload: Message<'CONTEXT_MENUS_CLICK'>['payload'] = {
     menuItemId
   }
-  try {
-    const tabs = await browser.tabs.query({ active: true, currentWindow: true })
-    const tab = tabs[0]
-    if (tab && tab.url) {
-      payload.linkUrl = tab.url
-      if (tab.id) {
-        try {
-          const word = await message.send<'PRELOAD_SELECTION'>(tab.id, {
-            type: 'PRELOAD_SELECTION'
-          })
-          if (word && word.text) {
-            payload.selectionText = word.text
-          }
-        } catch (e) {
-          /* */
+
+  const tabs = await browser.tabs
+    .query({ active: true, currentWindow: true })
+    .catch((): browser.tabs.Tab[] => [])
+
+  const tab = tabs[0]
+
+  if (tab && tab.url) {
+    payload.linkUrl = tab.url
+    if (tab.id) {
+      try {
+        const word = await message.send<'PRELOAD_SELECTION'>(tab.id, {
+          type: 'PRELOAD_SELECTION'
+        })
+        if (word && word.text) {
+          payload.selectionText = word.text
         }
+      } catch (e) {
+        console.error(e)
       }
     }
-  } catch (e) {
-    /* */
   }
-  message.send({
+
+  await message.send({
     type: 'CONTEXT_MENUS_CLICK',
     payload
   })