+++ test/specs/components/dictionaries/googledict/fixtures.js
@@ -2,16 +2,20 @@ module.exports = {
   files: [
     [
       'chs-mouse.html',
-      'https://www.google.com/search?hl=en&safe=off&q=define:mouse'
+      'https://www.google.com/search?hl=en&safe=off&q=meaning:mouse'
     ],
     [
       'chs-爱.html',
-      'https://www.google.com/search?hl=en&safe=off&q=define:' +
+      'https://www.google.com/search?hl=en&safe=off&q=meaning:' +
         encodeURIComponent('爱')
     ],
     [
       'en-love.html',
-      'https://www.google.com/search?hl=en&safe=off&q=define:love'
+      'https://www.google.com/search?hl=en&safe=off&q=meaning:love'
+    ],
+    [
+      'en-salad.html',
+      'https://www.google.com/search?hl=en&safe=off&q=meaning:salad'
     ]
   ]
 }